Changed color scheme slightly
The previous color scheme of blending from orange to pink across some of the elements was cringeworthy. Now I'm blending from orange to yellow which kinda reflects the beach. Combined with the purple/blue blends on other elements, I think this works quite nicely. The site looks bright and happy.